CRUD operations were modified

+++ Library.Application/Features/Authors/Commands/DeleteAuthor.cs
@@ -1,16 +1,19 @@
 ﻿using AutoMapper;
+using Library.Application.Common.Mappings;
+using Library.Application.DTOs;
 using Library.Application.Interfaces.Repositories;
-using Library.Domain.BaseEntities;
+using Library.Domain.Common;
 using Library.Domain.Entities;
 using Library.Shared.Results;
 using MediatR;
 
 namespace Library.Application.Features.Authors.Commands
 {
-	internal class DeleteAuthorCommand : BaseAuthor, IRequest<Result<int>>
+	public class DeleteAuthorCommand : BaseAuthor, IRequest<Result<int>>, IMapTo<Author>
 	{
+		public int Id { get; set; }
 	}
-	internal class DeleteAuthorHandler : IRequestHandler<DeleteAuthorCommand, Result<int>>
+	public class DeleteAuthorHandler : IRequestHandler<DeleteAuthorCommand, Result<int>>
 	{
 		private IUnitOfWork _unitOfWork;
 		private IMapper _mapper;
@@ -19,12 +22,16 @@ namespace Library.Application.Features.Authors.Commands
 			_unitOfWork = unitOfWork;
 			_mapper = mapper;
 		}
-
+		bool IsEqual(Author fromDb, AddAuthorCommand fromRequest) => _mapper.Map<AuthorDTO>(fromDb) == _mapper.Map<AuthorDTO>(fromRequest);
 		public async Task<Result<int>> Handle(DeleteAuthorCommand request, CancellationToken cancellationToken)
 		{
 			var author = _mapper.Map<Author>(request);
+			var authorFromDb = await _unitOfWork.Repository<Author>().GetByIdAsync(request.Id);
+			if (authorFromDb is null)
+				return await Result<int>.FailureAsync("Author doesn't exist");
 			await _unitOfWork.Repository<Author>().DeleteAsync(author);
-			return Result<int>.Success($"{author.Name} {author.Surname} was deleted successfully");
+			await _unitOfWork.Save();
+			return Result<int>.Success(request.Id, $"{author.Name} {author.Surname} was deleted successfully");
 		}
 	}
 }
